Move common functionality in a shared base class.
Update affected code.
